Move Parallel_callback out of function + documentation
